Convert Congress/Session tables to new structure
